+++ CHANGES.md
@@ -1,4 +1,6 @@
 #### [unreleased]
+
+#### 2.3.0 / 2019-01-22
 * add bypass for WordPress 5.1 WSOD Shutdown Handler
 * move loading hooks out of constructors
 * pass saved options to class constructors